Make up-to-date detection work for AsciidoctorTask
After it's been run once, Gradle considers AsciidoctorTask to be
up-to-date even if a file beneath sourceDir has been changed. Update
sourceDir so that it is annotated @InputDirectory rather than @Input
so that all of the files beneath sourceDir are considered when Gradle
is determining whether or not the task is up-to-date.