Fixes second part of issue #83
The problem here is actually not with includes per se but
with Asciidoctor's handling of absolute paths in Windows.
They are apparently treated as URIs and rendered as links only, unless
allow-uri-read attribute was specified.

This patch works around this problem by setting projectdir and rootdir
attributes as relative paths instead of absolute paths.