Ensure that AsciidoctorTask is not up-to-date when classpath changes
Previously, the AsciidoctorTask did not consider its classpath to be
an input. This meant that a change to task's classpath did not stop
it from considering itself to be up-to-date.

This commit annotates the classpath property with @InputFiles so that
it's considered when performing up-to-date checks. It's also been
annotated with @Optional as the property can be null.