Legacy attributes deprecation message triggers selectively (#270)
Potential source documetns are now scanned for the usage of
projectdir and rootdir. A warning message will only be printed when found.

If the Gradle vesion is < 4.5 the list of files will always be printed.
In later versions the additional files will only be printed if the
`--warning-mode=all` is set on the command line.