Additional attribute providers can be registered (#277)
The AsciidoctorJ extension can be used to register additional providers of
attributes of a global scale. Each Asciidoctor task can reuse these or
override with a custom list. Attributes provided via these registered
providers do no affect the up-to-date status of the tasks.

Default attributes are postifixed with a `@` when sent to Asciidoctor.
This allows documents to override default attributes in the case of a
buildscript author not having explcitly set those attributes.