Allows simplified strategies for controlling the base directory (#369)
Even with standardisation in 2.0, there are use cases which confuse users
or surprises their contextual expectations. Therefore it is now possible
to use specific strategies for controlling the base directory:

- Base directory is root project directory
- Base directory is project directtory
- Base directory follows the source directory even if an intermediate
  working directory is being used.

By default the base directory will still be the project directory as has
been the case for 1.5 & 2.0. The base directory can also be set to be any
arbitrary directory as required.