Source trees should report all language source sets (#401)
`getSourceFileTree` and `getSecondarySourceFileTree` now reports all of the source sets when
multiple languages are specified. This fix distinguishes between classic behaviour and
multi-language behaviour and will either return a single source set of a combination of source sets.

Credits to @ixchelruiz for helping to diagnose this problem at Hack.Commit.Push 2019.