Exlude secondary source elements by matching rather than substraction (#410)
This is a very subtle issue. When two Filtree instances are subtracted from each other
in order to eliminate a set of files, the base directory of the tree is lost. If this is
in a CopySpec the relative paths will thus also be lost. The solution is to use a matcher
of type PatternFilterable and then to exclude by FileTreeElement.

When using