Allow actions to be executed when Asciidoctorj detached configuration is created (#511)
Build script authors can registers additional actions to be executed every time the configuration
that holds the asciidcotrj dependencies, is created. This is done via the onConfiguration call.
Actions can be task-specific and can even be completely decoupled from the project actions. If not,
then project actions will be executed before task-specific actions.