Add support for giving multiple font directories for pdf
It was not possible to specify multiple directories when setting a custom pdf
font. If user specified paths with file separator it was treated as a single
directory.

New tests were added and the existing ones are updated.

With these commits, user can now specify multiple paths for the pdf font.