(#756) Generic output formatters
Added support to both asciidoctorj and asciidoctor.js for adding
output formatters that are not directly supported by this plugin
suite.

The asciidoctorj formatter can add JARs to the classpath or add
GEMs to the appropriate GEM configuration so that they can be
prepared.

The asciidoctor.js formatter can add NPM packages.